Implementação de melhorias no código
