+++ src/models/__init__.py
@@ -1,2 +1,1 @@
 from .models import abstracts, author, chairs, edition, paper, paperAuthor, reference, citation
-from .converters import convert_to_csv
